Refactor: Kafka Avro Tools
